+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.10.0
+
+### Minor Changes
+
+- updated errors-nest to renamed baseError to baseApiError
+
 ## 0.9.1
 
 ### Patch Changes